fixed so const ref only is used when passing larger structures.

+++ Solution/Game/EntityFactory.cpp
@@ -203,7 +203,7 @@ void EntityFactory::LoadBulletComponent(EntityData& aEntityToAddTo, XMLReader& a
 
 			aDocument.ForceReadAttribute(e, "value", damage);
 
-			aEntityToAddTo.myDamage = static_cast<unsigned short>(damage);
+			aEntityToAddTo.myDamage = damage;
 		}
 	}
 }
@@ -294,7 +294,7 @@ void EntityFactory::LoadHealthComponent(EntityData& aEntityToAddTo, XMLReader& a
 
 			aDocument.ForceReadAttribute(e, "value", life);
 
-			aEntityToAddTo.myLife = static_cast<unsigned short>(life);
+			aEntityToAddTo.myLife = life;
 		}
 	}
 }
@@ -309,7 +309,7 @@ void EntityFactory::LoadPhysicsComponent(EntityData& aEntityToAddTo, XMLReader& 
 		{
 			float weight = 0;
 			aDocument.ForceReadAttribute(e, "value", weight);
-			aEntityToAddTo.myEntity->GetComponent<PhysicsComponent>()->Init(weight);
+			aEntityToAddTo.myEntity->GetComponent<PhysicsComponent>()->Init(static_cast<unsigned short>(weight));
 		}
 	}
 }
